+++ projects/MOCOV3/train.sh
@@ -9,6 +9,6 @@ PORT=${PORT:-12345}
 
 python3 -m oneflow.distributed.launch \
 --nproc_per_node $GPUS --nnodes $NODE --node_rank $NODE_RANK --master_addr $ADDR --master_port $PORT \
-projects/MoCo_v3_Vit/train_net.py \
+projects/MOCOV3/train_net.py \
 --config-file $CONFIG ${@:3}
